Run a mine update before any of the orchestration begins
This is related to bug https://github.com/saltstack/salt/issues/11501,
but not limited to it.

This is happening because we aren't starting the salt-minion with an
already defined set of roles (roles in /etc/salt/grains). We are setting
roles dynamically through the dashboard, and after doing this, mine_functions
are not executed immediately. This means that we might be lacking the IP
address for the etcd node, for example, making other services reference a
blank IP address.

By running a minion.update before anything else we ensure that all nodes
will run their mine_functions, updating the mine correctly, and then
executing the highstate on every node with their information updated.