Update default etcd cluster size to match number of masters
Update the etcd role to dynamically choose the cluster size by default, using
the number of masters as the value. This dynamic value may still be overridden
by a pillar value.

We intentially don't make any attempts to prevent "invalid" cluster sizes here,
as doing so would cause issues during master addtion and removal (when that
becomes a suppored operation).