Update etcd member count logic
When etcd discovery is used, and the expected member count is larger
than the available pool of nodes, etcd will fail to start. So, we add
some logic to etcd_discovery_helpers to ensure we don't attempt to
form a cluster that is larger than the available pool of nodes.