Improve bootstrapping performance
* Do not generate a minion certificate on every machine that needs to
  download the CA certificate.
* Generate certificates per service, not per machine.

The rationale behind this is to generate keys/certificates per services for now,
since they are way easier to handle, and it will happen once at the beginning
of the orchestration, and then the certificates/keys will be deployed to each
machine depending on its role. At the very beginning of the orchestration three
certificates will be generated:

- CA certificate (authority, key kept always private)
- etcd certificate (published on the mine)
- apiserver certificate (published on the mine)

After the last two have been generated and created will be published on the mine,
and we will use it for deploying them on the rest of the minions depending on the
role of the machine.

This also allows us for a much better way of handling certificate expiration. We
would update the certificate, publish the new certificate on the mine, and then
deploy the new certificates like we do normally (and all related services would be
automatically restarted as required).

Split the phase of generation and installation of certificates, since they are
clearly two different steps.

CC: bsc#1038814