Make cert always include `fqdn`
The only component that was adding `fqdn` to the list of dns names of SAN
certificates is the `kube-master` role.

However, depending on the size of the cluster and other possible reasons
it might happen that a etcd member falls in a `kube-minion` instance,
where the certificate is missing local ip addresses, as well as the `fqdn`
of the machine. With this change, we are enforcing `cert` to always
generate this information automatically, while we still allow to extend
it, in case that's still necessary (for example, as kubernetes-master still
requires).

Check https://bugzilla.novell.com/show_bug.cgi?id=1039269#c9 for further
information.

Fixes: bsc#1039269