Stop managing /var/lib/etcd in salt
This dir is created by the etcd rpm, and permissions
are maintained by etcd when it is running

The salt and etcd disagree an what these permissions are
causing extra "changed" entries. As etcd is changing
them to what it needs, and the directory is created
by etcd (and its RPM) we should not try and manage it.