salt-api: listen to localhost [bsc#1043589]
Do not expose the salt-api to the entire world. This is needed only by
Velum to trigger salt actions. Given both the containers use the same
network namespace we can just bind this service to localhost.

By doing that we are going to reduce the attack surface.

This fixes one of the two issues reported by bsc#1043589

Signed-off-by: Flavio Castelli <fcastelli@suse.com>