Add support for initializing certificates (user certificates)
We add a reactor that will run an orchestration that will be executed
during the first boot of the services.

This new orchestration will make sure the CA generates both a private
and a public CA certificates, and depending on the service one or the
other will be used. E.g. salt-api uses the private CA, while velum will
use the public CA.

In the future, this salt-states can be modified, so the public CA is
a CA provided by the customer.

Fixes: bsc#1043589