Increase salt-master timeout
When dealing with a large number of minions, timeouts are visible when
using the default value of 5 seconds. Increasing the CPU/RAM resources
allocated to the master helps, but given it it's short bursts of heavy usage
(bootstrap and upgrade), this shouldn't be necessary.

We increase the timeout from 5 to 20 seconds, allowing tasks to take longer
yet still succeed.