Handle curl proxy settings
YaST is also configuring proxy settings inside of `/root/.curlrc`,
this is needed because zypper is using libcurl. So if you run zypper
from a cronjob or `su`, the `/etc/sysconfig/proxy` variables are not parsed
and set in the environment. Which means, zypper will not use the proxy
and fail. With `/root/.curlrc`, libcurl will use the proxies configured
there.

Signed-off-by: Flavio Castelli <fcastelli@suse.com>