Remove unneeded signing policies
These signing policies were used when the CA wasn't containerized,
when we containerized it, they were moved to `caasp-container-manifests`,
and the CA container is mounting it from there.

If we uncontainerize the CA in the future we can move it back if
needed, but let's keep this clean so it's not misleading.