Explicitly refresh the mine on all minions after the `ca` has published the `ca.crt`
We will explicitly force all minions to refresh the mine after the `ca` minion has
published the `ca.crt` certificate on the mine, to avoid rendering problems with
later SLS being executed. It might happen that a minion was missing this information
on its mine, so the rendering of the SLS failed, effectively stopping the whole
orchestration process.

Fixes: bsc#1048548