Use haproxy to load balance Kube API requests
Now that we can have multiple masters, we need a way for the various
services and end-users to be load balanced over the set of kube-api
servers.

We install haproxy on each node, inside a docker container, configured
to load balance requests over all the cluster masters. This haproxy
is configured to listen on 0.0.0.0 on the masters, and 127.0.0.1 on
the workers.

This is to allow the minions to simply "talk" to 127.0.0.0, and be
routed to an active kube-api server.