Use apply instead of create for addons
kubectl apply is generally idempotent, while kubectl create is not. With
multi-master now enabled, if two masters execute this script at once, one of
them is likely to fail given the check+set race within this script - Switching
to apply removes part of this this C+S race.

The second part of this race, is it client-side decision by apply to create or
update, by retrying the command once if it fails, we can ensure when two masters
run this script at the same time, for the first time, the C+S race will be
avoided here too.