Move External FQDN to 127.0.0.1 address
s was added to ensure Dex was always reachable, however, with multi masters,
this name was assigned to 3 different lines in /etc/hosts. Most consumers of
/etc/hosts do not deal with this as they would a round-robin DNS entry which
returns multiple IPs.

When the "selected" master is powered off, this name continues to resolve
the same dead IP address. As Dex uses a NodePort service, putting this to
127.0.0.1 works as we expect it to.