Making the service account key the same on all nodes
The kube-apiserver and kube-controller-manager must agree on what the
private key is for service account generation. In a multi-master
scenario, where an api server starts on one machine, and the
controller-manager on another machine becomes primary, pods cannot be
created because kube-controller-manager cannot communicate with the
apiserver.

So, now, we generate the service account key on the ca minion and store
it in the mine, so that it's generated once.

This will break if/when there is more than one machine with the ca role
in Salt.