Add a delay before retrying kubectl commands
We currently retry straight away, which can sometimes simply be too fast and we'll
get a similar error again. Add a short 3 second delay before retrying any kubectl
commands.

Additionally, remove the check_cmd's from the K8S addons, as these don't actually
check anything which has not already been asserted through the primary command
itself, and again, these occasionally run too fast, before K8S has fully processed
the tasks from the primary command.

For more easily finding info related to the pattern used here in git history, I'm
including a comment on this pattern here:

This pattern is used in at least 3 different places in the salt codebase (dex,
kube-dns, helm). The issue is that: kubectl apply makes a client side decision on
if it needs to issue a "CREATE" or "UPDATE" API call. When more than 1 master
executes this at the same time, for the first time, the check to decide "create
or update" returns "create" on more than 1 master, assuming the timing is right.

Now, when both send the create API call, kubernetes will reject one of them with
an "AlreadyExists" failure. This will cause salt to fail the bootstrap. However,
if we try the command again on the machine that failed, it will this time see the
resource created by the other master, and decide to issue a no-op UPDATE instead.
This will pass, and all resources will be in a consistent state - at the expense
of a an extra few API calls.

bsc#1062818