Revert K8S to use etcd2 storage format
With etcd3, the kubernetes api server will sit in a (slow) restart loop
when multimaster is enabled, logging a stacktrace and then restarting.
This will manifest as, most commonly, "Unable to connect to the server:
unexpected EOF" from kubectl. This will break bootstrap as we need to
talk to K8S API to deploy dex, kube-dns, and tiller.

bsc#1063235
bsc#1063285
bsc#1063543