helm should detect salt-installed tiller service
The helm client looks for a tiller deployment called 'tiller-deploy' to
establish if tiller is already installed in the cluster, or not. Update
our salt install of tiller to use a deployment with the same name so
that it will be recognized by the helm client as already being
installed.

Fixes: bsc#1066201