Migrate CNI metadata on workers before doing anything else
This does not give any chance for kubelets to try to request a new `podCIDR`.

Also, fix node patching of the CNI migration

Before restarting the master with the new configuration we migrate the workers
to their expected `podCIDR` values, then we start with the general update
procedure: masters first, then workers.