Remove iptables rules.
The cluster seems to work fine without this rules, and also we are
not persisting them, so they are lost after a reboot.

Try to keep the deployment as simple as possible and only do the
actions we require to work.

(cherry picked from commit 8339f6f3b06eb9f6bf1afc167c01e820b5557735)