Namespace the roles and cluster roles we create
When we create a role, rolebinding etc, we should namespace the names
in order to make it obvious these are deployed as part of CaaSP, as well
as to help ensure these are obviously part of CaaSP, not a stock part of
Kubernetes.

I've gone with a "suse:caasp:" prefix, which matches the "system:" prefix
for built in roles/rolebindings/etc.