Mark the haproxy as critical pod
Flag the haproxy pods providing connectivity to the API server as critical ones.

This should force kubelet and the scheduler to never ever get rid of
them. If these pods are killed to make more space for other ones, the
node would not be able to talk with the API server making it useless.

More details inside upstream doc:
https://kubernetes.io/docs/tasks/administer-cluster/guaranteed-scheduling-critical-addon-pods/

Signed-off-by: Flavio Castelli <fcastelli@suse.com>