When executing a highstate of `apiserver` make sure that we check the local `apiserver` instance
When executing the highstate make sure the `apiserver` we are checking
is the local one, not *any* master through haproxy.

Make haproxy more reliable.

- Let it redispatch requests.
- Really restart the service when the config changes.
- Apply configuration before highstates with a small batch, so we control the restarts.

Wait for the apiserver to be up and responding behind HAProxy

Remove `addons` and `dex` as a trait of the `kube-master` role.

Instead, deploy them as part of the orchestration.

Fixes: bsc#1079460