Don't Merge, Won't work, Just an example
An example of how we might split out the phases within each salt module, allowing
us to be more targetted in what we call and when - e.g. when velum wants to reconfigure
docker - the reconfig orch can just call the parts it needs, and skip the rest - e.g:

* docker.configure
* docker.reload

However - this won't work as is, as all the requires/watches/etc are gone. Those need
working out...