Update `etcd` certificates before updating any machine
We need to include the new SAN on all the certificates before restarting
the first machine. Otherwise, this machine (a master) can find itself
isolated without being able to contact any etcd member with the name it
has (as the rest of the nodes haven't updated their certificates yet to
also include the new name on the SAN).