When populating the cache, don't fail if this fails for some reason.
There's a race condition in which the cache directory does not exist, but
when tried to be created it has already been created by something else, and
an exception is raised, stopping the execution.

When populating the cache, we don't really care if it was correctly populated
or not in that *specific* call, so move on.

Fixes: bsc#1084441