Ensure Dex certificate rotations work as expected
When the Dex config / certs are changed, nothing would trigger
the restart of the already running Dex pods in order to have Dex
load it's new config, or certificates.

A common pattern for this is to use checksum annotations on the pods,
when the annotation changes, Kubernetes will start the rolling upgrade
of the pods - restarting them according to their rolling upgrade
policy.

Fixes bsc#1082996