When upgrading, make sure to refresh the grains before attempting a highstate
During an upgrade process, some important grains (like `nodename` during certain
upgrades, others in the future could also apply) might be outdated when we
highstate the node. Usually the salt-minion does this automatically, but QA has
found some cases in which this doesn't happen.

This is an attempt to fix that issue, despite hardly reproducible.

feature: upgrade