Fix issue caused by velum pillar override
Pillars set by velum are going to override what is set via the
`salt/pillars` files.

That caused all the nodes to be using cri-o. The following code enforces
'docker' to be used for all the nodes with a certain role (eg: the admin
and the ca ones).

feature#crio

Signed-off-by: Flavio Castelli <fcastelli@suse.com>