Remove unused cri abstractions
cri-o doesn't have yet a way to copy files from the host into its running
containers. Fortunately this feature is required only on the admin node,
which is still using docker.

This commit removes some of the abstractions introduced to be able to
copy files into running containers.

We will revert this commit later on, once we migrate the admin node to
use cri-o.

feature#crio

Signed-off-by: Flavio Castelli <fcastelli@suse.com>