Always try to remove the etcd member
Don't try to be smart here: always try to unregister the etcd member,
since `manage.up` and `manage.down` won't list the down node using
roles, as salt cannot access it.

Just try to always remove the etcd member, if it's not an etcd member
it's fine, nothing will happen.

Fixes: bsc#1090392