kubelet: allow resource reservation
Allow kubelet to take into account resource reservation and eviction
threshold.

== Resource reservation ==

It's possible to reserve resources for the `kube` and the `system`
components.

The `kube` component is the one including the kubernetes components:
api server, controller manager, scheduler, proxy, kubelet and the container
engine components (docker, containerd, cri-o, runc).

The `system` component is the `system.slice`, basically all the system
services: sshd, cron, logrotate,...

By default don't specify any kind of resource reservation. Note well:
when the resource reservations are in place kubelet will reduce the
amount or resources allocatable by the node. However **no** enforcement
will be done neither on the `kube.slice` nor on the `system.slice`.

This is not happening because:

  * Resource enforcement is done using cgroups.
  * The slices are created by systemd.
  * systemd doesn't manage all the available cgroups yet.
  * kubelet tries to manage cgroups that are not handled by systemd,
    resulting in the kubelet failing at startup.
  * Changing the cgroup driver to `systemd` doesn't fix the issue.

Moreover enforcing limits on the `system` and the `kube` slices can lead
to resource starvation of core components of the system. As advised even by
the official kubernetes docs, this is something that only expert users
should do only after extensive profiling of their nodes.

Finally, even if we wanted to enforce the limits, the right place would
be systemd (by tuning the slice settings).

For more information see the official documentation:
https://kubernetes.io/docs/tasks/administer-cluster/reserve-compute-resources/

== Eviction threshold ==

By default no eviction threshold is set.

bsc#1086185

Signed-off-by: Flavio Castelli <fcastelli@suse.com>