Lower the per-request timeout when we are checking for successful query
When we are waiting for some service to be up, if the request hangs for
some reason, we want to retry at least several times. Without setting
this value explicitly, it takes the default (`http_request_timeout` as 3600),
what is way over our `wait_for` argument set at 300 seconds.

By setting the default `http_request_timeout` to a more reasonable default
when doing this kind of checks we can ensure that the request itself will
timeout several times before we call it done.

Fixes: bsc#1093540
Fixes: bsc#1093685