Uncordon node in a explicit sls action
This way we don't try to uncordon the node in the `kubelet/init.sls` file,
required for example by `haproxy`, that will end up in the machine trying
to early uncordon itself (when `haproxy` configuration hasn't been written
yet, and leading to early failure).

Splitting this action and called only when required (this is: the update
process) is safer.

Fixes: bsc#1080978