Remove mine information when removing a node
This will avoid to render stale information about critical components,
like `etcd` endpoints in the `etcd` configuration.

`etcd` is very sensitive to this kind of misleading (stale) information,
if more endpoints are provided in `ETCD_INITIAL_CLUSTER` than the ones
that actually exist in the cluster, a new instance of etcd will refuse
to start.

Fixes: bsc#1097001
Fixes: bsc#1097147