Salt velum pillar: send minion_id
Provide the minion_id requesting the pillars to the Velum endpoint, this
allows Velum to filter results.

This is required by the AdmissionWebhook feature to avoid to send the
API server client certificate and key to nodes that are **not** master
nodes. Other nodes don't need these confidential information.

feature#ValidatingAdmissionWebhook

Signed-off-by: Flavio Castelli <fcastelli@suse.com>