Allow `etcd` to grow as required and shrink to optimal etcd cluster sizes on corner cases.
Improve `etcd` configuration handling to allow it to grow as needed. This
change includes:

* Adding several masters at the same time
  ** `etcd` will grow instance by instance still, as recommended by the
     `etcd` administration best practices.

* Try to use the current endpoints reported by `etcd`. This makes much
  easier to grow several instances one by one without having to relay
  on internal hacks to properly set up `ETCD_INITIAL_CLUSTER` environment
  variable.

* Add helper methods that allow us to list current members (active and
  unstarted)

* Differentiate between the first bootstrap (`ETCD_INITIAL_CLUSTER_STATE`
  defaults to `new`) and *any* other run, where `ETCD_INITIAL_CLUSTER_STATE`
  will be `existing`, as the `etcd` cluster is already running.

When we grow, we take into account the golden ratio; however, when shrinking
the cluster we don't. It might happen that a cluster ends up with not
recommended etcd number of nodes (2, 4, 6...) depending on how it grew before
and how it shrank.

This logic makes sure that we are always on an etcd golden ratio, also
on corner cases when removing nodes.

Fixes: bsc#1098433
Fixes: bsc#1098064
Fixes: bsc#1098161
(cherry picked from commit 6b86e9e725d0f5ff6f63c5257604e13499505482)