During upgrade, ensure masters always have the correct taints
When migrating from the "old" to "new" names for the kubelets, we pre-create
the new node so that we can clone the network config. This means the kubelet
is NOT self-registering, and the "single use options" like --register-with-taints
are ignored. This means the kubelet is connected from the period of time where it
starts, to where salt later forcefully adds the taint. Any pods created during
this window could end up scheduled to the master.

bsc#1098383