Call to `mine.update` after `saltutil.sync_pillar` has been called.
During an upgrade we want to call to `mine.update` after `saltutil.sync_pillar`
has been called, because the `mine_functions` reside on the pillar, we first want
to make sure to sync that, and update the mine afterwards. Otherwise, we risk
doing this in a race condition when the salt minion starts, and it could or
could not lead to update orchestration failure.

Fixes: bsc#1097478