Bypass haproxy and wait for the kubelet to be back before continuing with the upgrade process.
During the upgrade phase from 2.0 to 3.0 the authentication changed on the cluster. This means
that the worker nodes won't be able to authenticate with any master once that all masters have
been updated. However, when updating we need to drain the workers. Otherwise we risk to hit
a bug that makes the update hang; if a worker node has active ceph mounts, systemd will hang
upon reboot.

After all masters have been updated, we'll make all workers bypass their local haproxy (it is
not yet updated to forward auth information to the apiserver), and they can contact the haproxy
instance on one of the masters directly; this way we can successfully drain them during the
upgrade process.

Fixes: bsc#1100212