Add haproxy migration sls to apply during upgrade
During an upgrade from 2.0 to 3.0, workers will lose communication with
the apiservers on the master nodes because of an auth change. After we
have applied all the master nodes, and before we start looping over the
workers, apply haproxy system-wide on all the workers, allowing their
haproxy to update its configuration, thus, being able to authenticate
against the apiservers again.

This patch includes a new tree structure, meant to be destroyed between
versions, but that allows to not poison the main structure of states
with transient migration logic. The structure is as follows:

- migrations
  - <orig_version>-<target_version>
    - overriden-sls/*
    - * (direct actions that can spawn other migration tasks)

Fixes: bsc#1100212