Pass `ETCDCTL_API` version as an envvar instead of at the command name.
This would not work as what is fork + exec'ed is the `ETCDCTL_API=version`,
what is not found in the `PATH`, causing the whole command to fail.

Besides, each version produces a different output, and the parsing needs to
be updated in order to properly get the member id.

Follow up of https://github.com/kubic-project/salt/pull/622