Only add a new etcd member if no alias is already a member
When adding a new member to etcd, it might happen that it is already
part of the cluster using one of the aliases - when migrating from
v2 to v3 it seems common that the default nodename changes.

If this is the case it should not be added again with the new nodename,
as one node can not have 2 etcd members.