Fix 2 to 3 upgrades when draining nodes before rebooting
Before rebooting a 2 machine we drain it; make sure that when we send
the drain command in this case it's sent to the local apiserver running
in this machine, and only do this in the specific case in which the machine
is in the 2.0 version. Otherwise, perform the general case.

Fixes: bsc#1112744