Enable kube-apiserver authentication to the kubelet (bsc#1121146)
Kube-apiserver should authenticate to the kubelet with a client certificate and key.
This is configured with the --kubelet-client-certificate and --kubelet-client-key
flags provided to the API server. Kubelet has to be started with the --client-ca-file
flag or clientCAFile option in kubelet-config.yaml file, this is providing a CA bundle
to verify client certificates with.

(cherry picked from commit 6309fb22ae122db6e2db2705fe47c1f4ae939ffb)