velum: set x-frame-options deny in haproxy cfg
The HTTP Header X-Frame-Options provides a defense against framing
attacks, whereby an attacker causes a target page to be loaded within
an IFrame or similar. An example of this type of attack is
Click-jacking.

With X-Frame-Options set to SAMEORIGIN it is possible that any other
page within the same domain, if compromised, could be used to perform
such attacks.

To avoid this we set that to DENY since so far we don't need any
iframe.

Signed-off-by: Vítor Avelino vavelino@suse.com

bsc#1121346

(cherry picked from commit ffc3cf9092a4d697bde19cba0cb8170334efe6d7)