Jenkinsfile: Use docker cmdline directly instead of k8s Jenkins plugin
The tox and flake8 pipelines are the only ones which depend on the k8s
Jenkins plugin. As such, we can use docker directly in order to be able
to drop the plugin from the server. The nodelabel is hardcoded because it
does not make much sense to make this configurable given everything happens
on a container.

(cherry picked from commit 8443e0d863790cb98ba09fd5e79e1df28dd9d1e5)