Jenkinsfile: Use docker cmdline directly instead of k8s Jenkins plugin
The tox and flake8 pipelines are the only ones which depend on the k8s
Jenkins plugin. As such, we can use docker directly in order to be able
to drop the plugin from the server. The nodelabel is hardcoded because
it does not make much sense to make this configurable given everything
happens on a container.

(This is a backport for kubic-project/salt#722 since we are not supposed
to commit stuff to the public release-3.0 branch anymore)