Make nodename appear first on the /etc/hosts file
Salt will pick the first name on the current default interface to determine
the hostname of the machine. Since we are sorting with all entries for each
machine there's a high change that a salt minion id will win the first position,
affecting certain grains that we use to determine the hostname of the node.

Fixes: bsc#1117339
(cherry picked from commit 4f75ad33fdffc3ea88c91c56e56fb5af9275714d)