Jenkinsfile: Switch to dynamic library fetching and drop branch
Instead of having the library hardcoded to Jenkins master, we can fetch
it dynamically. We also drop the usage of library branches since it does
not make sense to maintain such a thing in the CI. The master branch
should be able to handle both development and release branches.

Backport of kubic-project/salt#741 and kubic-project/salt#744