Automatically update the kubernetes-service endpoint.
When a cluster is bootstrapped with multiple kube-apiservers, the `kubernetes`
service contains a list of all of these endpoints.

By default, this list of endpoints will *not* be updated if one of the
apiservers goes down. This can lead to the api becoming unresponsive and
breaking it. To have the endpoints automatically keep track of the apiservers
that are available the `--endpoint-reconciler-type` option `lease` needs to be
added.

(The default option for 1.10 `master-count` only changes the endpoint when the
count changes: https://github.com/apprenda/kismatic/issues/987)

See:

https://github.com/kubernetes/kubernetes/issues/22609
https://github.com/kubernetes/kubernetes/issues/56584
https://github.com/kubernetes/kubernetes/pull/51698